Crate substrate_executor

Temporary crate for contracts implementations.

This will be replaced with WASM contracts stored on-chain. ** NOTE *** This is entirely deprecated with the idea of a single-module Wasm module for state transition. The dispatch table should be replaced with the specific functions needed:

execute_block(bytes)

init_block(PrevBlock?) -> InProgressBlock

add_transaction(InProgressBlock) -> InProgressBlock It is left as is for now as it might be removed before this is ever done.

Re-exports

pub use wasmi;

Modules

error
Rust executor possible errors.

Macros

convert_args
Converts arguments into respective WASM types.

gen_signature
Generates a WASM signature for given list of parameters.

impl_function_executor
Implements wasmi::Externals trait and Resolver for given struct.

marshall
Pass the list of parameters by converting them to respective WASM types.

native_executor_instance
Implements a NativeExecutionDispatch for provided parameters.

unmarshall_args
Converts the list of arguments coming from WASM into their native types.

Structs

NativeExecutor
A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

NativeVersion

RuntimeVersion
Runtime version. This should not be thought of as classic Semver (major/minor/tiny). This triplet have different semantics and mis-interpretation could cause problems. In particular: bug fixes should result in an increment of spec_version and possibly authoring_version, absolutely not impl_version since they change the semantics of the runtime.

WasmExecutor
Wasm rust executor for contracts.

Traits

Codec
Trait that allows zero-copy read/write of value-references to/from slices in LE format.

Externalities
Externalities: pinned to specific active address.

NativeExecutionDispatch
Delegate for dispatching a CodeExecutor call to native code.

RuntimeInfo
Provides runtime information.

Functions

with_native_environment
Set up the externalities and safe calling environment to execute calls to a native runtime.